Updates types and fixes whitespace
