+++ src/helpers/siwf.ts
@@ -1,10 +1,16 @@
 import { SignatureFn } from "../types";
-import {
-  CreateLoginSiwfResponseArguments,
-  SignInPayloads,
-  SiwfResponse,
-} from "../siwf-types";
+import { SiwfResponsePayload, SiwfResponse } from "@projectlibertylabs/siwf";
 import { encodedValueToSignature, userAddressToPublicKey } from "./utils";
+
+export interface CreateLoginSiwfResponseArguments
+  extends Record<string, unknown> {
+  domain: string;
+  uri: string;
+  version: string;
+  chainId: string;
+  nonce?: string;
+  issuedAt?: string;
+}
 
 export async function createLoginSiwfResponse(
   userAddress: string,
@@ -12,14 +18,13 @@ export async function createLoginSiwfResponse(
   payloadArguments: CreateLoginSiwfResponseArguments,
 ): Promise<SiwfResponse> {
   const loginCaip122 = `${payloadArguments.domain} wants you to sign in with your Frequency account:
-    frequency:${payloadArguments.chainId}:${userAddress}
-    
-    URI: ${payloadArguments.uri}
-    Version: ${payloadArguments.version}
-    Nonce: ${payloadArguments.nonce}
-    Chain ID: frequency:${payloadArguments.chainId}
-    Issued At: ${payloadArguments.issuedAt}
-  `;
+frequency:${payloadArguments.chainId}:${userAddress}
+
+URI: ${payloadArguments.uri}
+Version: ${payloadArguments.version}
+Nonce: ${payloadArguments.nonce}
+Chain ID: frequency:${payloadArguments.chainId}
+Issued At: ${payloadArguments.issuedAt}`;
 
   const encodedValue = await signatureFn({
     method: "personal_sign",
@@ -47,7 +52,7 @@ export async function createLoginSiwfResponse(
 
 export async function createSignInSiwfResponse(
   userAddress: string,
-  payloads: SignInPayloads,
+  payloads: SiwfResponsePayload[],
 ): Promise<SiwfResponse> {
   const userPublicKey = userAddressToPublicKey(userAddress);
   return {
